+++ bee-vote/tests/vote_context.rs
@@ -1,8 +1,7 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod mock;
-
+use bee_test::rand::transaction::rand_transaction_id;
 use bee_vote::{
     context::{VoteContextBuilder, VoteObject},
     opinion::{Opinion, Opinions},
@@ -10,7 +9,7 @@ use bee_vote::{
 
 #[test]
 fn is_finalized() {
-    let ctx = VoteContextBuilder::new(VoteObject::Conflict(mock::rand_transaction_id()))
+    let ctx = VoteContextBuilder::new(VoteObject::Conflict(rand_transaction_id()))
         .with_initial_opinions(Opinions::new(vec![Opinion::Like; 5]))
         .build()
         .unwrap();
@@ -20,7 +19,7 @@ fn is_finalized() {
 
 #[test]
 fn is_not_finalized() {
-    let ctx = VoteContextBuilder::new(VoteObject::Conflict(mock::rand_transaction_id()))
+    let ctx = VoteContextBuilder::new(VoteObject::Conflict(rand_transaction_id()))
         .with_initial_opinions(Opinions::new(vec![
             Opinion::Like,
             Opinion::Like,
@@ -36,14 +35,14 @@ fn is_not_finalized() {
 
 #[test]
 fn last_opinion() {
-    let ctx = VoteContextBuilder::new(VoteObject::Conflict(mock::rand_transaction_id()))
+    let ctx = VoteContextBuilder::new(VoteObject::Conflict(rand_transaction_id()))
         .with_initial_opinions(Opinions::new(vec![Opinion::Like; 4]))
         .build()
         .unwrap();
 
     assert_eq!(ctx.last_opinion(), Some(Opinion::Like));
 
-    let ctx = VoteContextBuilder::new(VoteObject::Conflict(mock::rand_transaction_id()))
+    let ctx = VoteContextBuilder::new(VoteObject::Conflict(rand_transaction_id()))
         .with_initial_opinions(Opinions::new(vec![
             Opinion::Like,
             Opinion::Like,
